+++ src/GlobalHelper.php
@@ -176,9 +176,9 @@ class GlobalHelper
     {
         if (!isset(self::$cache)) {
             $home = PlatformUtil::getUserDirectory();
-            $directory = $home ? $home . '/.cache' : sys_get_temp_dir();
+            $directory = ($home ? $home . '/.cache' : sys_get_temp_dir()) . '/castor';
 
-            self::setCache(new FilesystemAdapter(directory: $directory . '/castor'));
+            self::setCache(new FilesystemAdapter(directory: $directory));
         }
     }