+++ tests/Examples/ContextContextDynamicTest.php
@@ -9,7 +9,7 @@ class ContextContextDynamicTest extends TaskTestCase
     // context:context
     public function test(): void
     {
-        $process = $this->runTask(['context:context', '--context', 'dynamic']);
+        $process = $this->runTask(['context:context', '--context', 'dynamic', '--no-trust']);
 
         $this->assertSame(0, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());