+++ tests/Examples/ContextContextProductionTest.php
@@ -9,7 +9,7 @@ class ContextContextProductionTest extends TaskTestCase
     // context:context
     public function test(): void
     {
-        $process = $this->runTask(['context:context', '--context', 'production']);
+        $process = $this->runTask(['context:context', '--context', 'production', '--no-trust']);
 
         $this->assertSame(0, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());