+++ tests/Examples/EnvEnvTest.php
@@ -9,7 +9,7 @@ class EnvEnvTest extends TaskTestCase
     // env:env
     public function test(): void
     {
-        $process = $this->runTask(['env:env']);
+        $process = $this->runTask(['env:env', '--no-trust']);
 
         $this->assertSame(0, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());