+++ tests/Examples/NoConfigUnknownWithArgsTest.php
@@ -9,7 +9,7 @@ class NoConfigUnknownWithArgsTest extends TaskTestCase
     // unknown:command
     public function test(): void
     {
-        $process = $this->runTask(['unknown:command', 'toto', '--foo', 1], '/tmp');
+        $process = $this->runTask(['unknown:command', 'toto', '--foo', 1, '--no-trust'], '/tmp');
 
         $this->assertSame(1, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());