+++ tests/Examples/ParallelSleepTest.php
@@ -9,7 +9,7 @@ class ParallelSleepTest extends TaskTestCase
     // parallel:sleep
     public function test(): void
     {
-        $process = $this->runTask(['parallel:sleep', '--sleep5', '0', '--sleep7', '0', '--sleep10', '0']);
+        $process = $this->runTask(['parallel:sleep', '--sleep5', '0', '--sleep7', '0', '--sleep10', '0', '--no-trust']);
 
         $this->assertSame(0, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());