+++ tests/Examples/RunTestFileTest.php
@@ -9,7 +9,7 @@ class RunTestFileTest extends TaskTestCase
     // run:test-file
     public function test(): void
     {
-        $process = $this->runTask(['run:test-file']);
+        $process = $this->runTask(['run:test-file', '--no-trust']);
 
         $this->assertSame(1, $process->getExitCode());
         $this->assertStringEqualsFile(__FILE__ . '.output.txt', $process->getOutput());